book-service: т.к. сервисы повоторяются, то в данной ветке было решено написать и все сервисы. Случилась проблема с update, т.к. его чуть подольше писать.

+++ src/main/java/ru/dynamica/library/controller/BookController.java
@@ -34,7 +34,7 @@ public class BookController {
     @PatchMapping("/${id}")
     @ApiOperation("Редактирование информации о книге")
     public ResponseEntity<BookDto> updateBook(
-            @PathVariable int id,
+            @PathVariable Long id,
             @RequestParam String attribute,
             @RequestParam String newValue
     ) {
